requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.2.0
django-autoslug==1.9.9
django-cleanup==9.0.0
django-filter==22.1
djangorestframework==3.14.0
dotenv==0.9.9
drf-spectacular==0.28.0
fastuuid==0.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
litellm==1.77.1
Markdown==3.4.1
MarkupSafe==3.0.2
multidict==6.6.4
openai==1.107.2
packaging==25.0
pillow==10.4.0
propcache==0.3.2
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.4
rpds-py==0.27.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tiktoken==0.11.0
tokenizers==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wikipedia==1.4.0
yarl==1.20.1
zipp==3.23.0
